requirements.txt
pandas==2.3.3
openpyxl==3.1.5
openai==2.14.0
pydantic==2.12.5
tiktoken==0.12.0
python-dotenv==1.2.1